Move RageDisplay to display folder
