+++ src/Etterna/Screen/Others/ScreenDebugOverlay.cpp
@@ -12,7 +12,7 @@
 #include "Etterna/Models/Misc/PlayerState.h"
 #include "Etterna/Singletons/PrefsManager.h"
 #include "Etterna/Singletons/ProfileManager.h"
-#include "RageUtil/Graphics/RageDisplay.h"
+#include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "RageUtil/Misc/RageInput.h"
 #include "Core/Services/Locator.hpp"
 #include "RageUtil/Sound/RageSoundManager.h"
