+++ src/Etterna/Screen/Others/ScreenTestSound.cpp
@@ -1,6 +1,6 @@
 #include "Etterna/Globals/global.h"
 #include "Etterna/Models/Misc/InputEventPlus.h"
-#include "RageUtil/Graphics/RageDisplay.h"
+#include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Etterna/Singletons/ScreenManager.h"
 #include "ScreenTestSound.h"
