+++ src/RageUtil/Graphics/Display/RageDisplay_Null.h
