+++ src/arch/InputHandler/InputHandler_X11.cpp
@@ -4,7 +4,7 @@
 #include <array>
 
 #include "RageUtil/Utils/RageUtil.h"
-#include "RageUtil/Graphics/RageDisplay.h"
+#include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "Etterna/Singletons/InputFilter.h"
 #include "archutils/Unix/X11Helper.h"
