+++ src/arch/MovieTexture/MovieTexture_FFMpeg.cpp
@@ -1,7 +1,7 @@
 #include "Etterna/Globals/global.h"
 #include "MovieTexture_FFMpeg.h"
 
-#include "RageUtil/Graphics/RageDisplay.h"
+#include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "Core/Services/Locator.hpp"
 #include "RageUtil/Utils/RageUtil.h"
 #include "RageUtil/File/RageFile.h"
